# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Smart Airport Ride Pooling API
  description: |
    High-performance ride pooling backend that groups airport-bound passengers into shared cabs,
    minimizing travel deviation while respecting seat/luggage constraints.

    **Targets:** <300ms latency · 100 RPS · 10,000 concurrent users
  version: 1.0.0
  contact:
    name: Hintro API
  license:
    name: MIT
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /api/v1/cancel/{request_id}:
    post:
      tags: [Booking]
      summary: Cancel a ride request
      description: |
        Cancels a ride request (real-time cancellations).
        Only PENDING and MATCHED requests can be cancelled.
      operationId: cancelRide
      parameters:
        - name: request_id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: 2
      responses:
        '200':
          description: Cancellation successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CancelResult'
        '400':
          description: Invalid request_id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Ride request not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Already cancelled or non-cancellable state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    CancelResult:
      type: object
      required: [request_id]
      properties:
        request_id:
          type: integer
          format: int64
        previous_trip_id:
          type: integer
          format: int64
        trip_cancelled:
          type: boolean
        cab_freed:
          type: boolean
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string